refactor: use namespace for fields
