+++ inc/exception/comparisonexception.class.php
@@ -29,8 +29,11 @@
  * ---------------------------------------------------------------------
  */
 
+namespace GlpiPlugin\Formcreator\Exception;
+use RuntimeException;
+
 if (!defined('GLPI_ROOT')) {
    die("Sorry. You can't access this file directly");
 }
 
-class PluginFormcreatorComparisonException extends Exception{}
+class ComparisonException extends RuntimeException {}
