+++ inc/field/hiddenfield.class.php
@@ -29,7 +29,13 @@
  * ---------------------------------------------------------------------
  */
 
-class PluginFormcreatorHiddenField extends PluginFormcreatorField
+namespace GlpiPlugin\Formcreator\Field;
+
+use PluginFormcreatorField;
+use Html;
+use Toolbox;
+
+class HiddenField extends PluginFormcreatorField
 {
    public function isPrerequisites() {
       return true;
